work with ProjectService

+++ src/main/java/by/dytni/test/service/ProjectService.java
@@ -2,6 +2,7 @@ package by.dytni.test.service;
 
 import by.dytni.test.models.Project;
 import by.dytni.test.repository.ProjectRepository;
+import by.dytni.test.repository.RecordRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,19 +13,37 @@ import java.util.Optional;
 public class ProjectService {
     @Autowired
     private ProjectRepository projectRepository;
+    @Autowired
+    private RecordRepository recordRepository;
 
     public void save(Project project) {
         projectRepository.save(project);
     }
-    public Project findById(Long id) {
-        Optional<Project> project = projectRepository.findById(id);
-        return project.orElse(null);
+    public List<Project> findByName(String name) {
+        return List.of(projectRepository.findByName(name).orElse(new Project()));
     }
     public List<Project> findAll() {
         return projectRepository.findAll();
     }
+    public void update(Long id, String name, String description) {
+        Project project = projectRepository.findById(id).orElse(null);
+        if (project == null) {
+            return;
+        }
+        if (!(name == null || name.isEmpty())){
+            project.setName(name);
+        }
+        if (!(description == null || description.isEmpty())){
+            project.setDescription(description);
+        }
+        projectRepository.save(project);
+
+    }
     public void delete(Long id) {
-        projectRepository.deleteById(id);
+        if(projectRepository.existsById(id)) {
+            recordRepository.deleteAllByProjectId(id);
+            projectRepository.deleteById(id);
+        }
     }
 
 }
